refactor: Introduce new class FilterInstructions
